@@ -1,10 +1,11 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
-		<TargetFramework>net6.0</TargetFramework>
+		<TargetFramework>net8.0</TargetFramework>
 		<RootNamespace>TowerSoft.Repository.MicrosoftSql</RootNamespace>
+		<ImplicitUsings>enable</ImplicitUsings>
 		<GeneratePackageOnBuild>true</GeneratePackageOnBuild>
-		<Version>4.0.0</Version>
+		<Version>4.1.0</Version>
 		<Company>TowerSoft</Company>
 		<Authors>StrutTower</Authors>
 		<Product>TowerSoft.Repository</Product>
@@ -37,11 +38,8 @@
 	</ItemGroup>
 
 	<ItemGroup>
-		<PackageReference Include="System.Data.SqlClient" Version="4.8.6" />
-	</ItemGroup>
-
-	<ItemGroup>
-	  <ProjectReference Include="..\TowerSoft.Repository\TowerSoft.Repository.csproj" />
+	  <PackageReference Include="Microsoft.Data.SqlClient" Version="6.0.2" />
+	  <PackageReference Include="TowerSoft.Repository" Version="4.1.0" />
 	</ItemGroup>
 
 </Project>
