@@ -1,12 +1,13 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
 	<PropertyGroup>
-		<TargetFramework>net6.0</TargetFramework>
+		<TargetFramework>net8.0</TargetFramework>
 		<RootNamespace>TowerSoft.Repository</RootNamespace>
+		<ImplicitUsings>enable</ImplicitUsings>
 		<AssemblyName>TowerSoft.Repository</AssemblyName>
 		<GeneratePackageOnBuild>true</GeneratePackageOnBuild>
 		<PackageId>TowerSoft.Repository</PackageId>
-		<Version>4.0.0</Version>
+		<Version>4.1.0</Version>
 		<Authors>StrutTower</Authors>
 		<Product>TowerSoft Repository</Product>
 		<Company>TowerSoft</Company>
@@ -40,8 +41,8 @@
 	</ItemGroup>
 
 	<ItemGroup>
-		<PackageReference Include="Dapper" Version="2.1.35" />
-		<PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="6.0.4" />
+		<PackageReference Include="Dapper" Version="2.1.66" />
+		<PackageReference Include="Microsoft.Extensions.Logging.Abstractions" Version="9.0.6" />
 	</ItemGroup>
 
 </Project>
